+++ CHANGELOG.md
@@ -97,6 +97,7 @@
 
 ### v3.8.13 / 2022-04
 
+* [1178](https://github.com/Vanessa219/vditor/issues/1198) 调用 insertValue 后第一次用户输入没有触发 input 回调函数 `修复缺陷`
 * [1179](https://github.com/Vanessa219/vditor/pull/1179) 更新代码风格工具&包管理工具 `开发重构`
 
 ### v3.8.12 / 2022-03-03